Added NavBar component Updated UI for Footer
To Do's

Add a filter component that pops up when the filter icon is clicked.
Add a filter state.
Create functions that update the API request based on the filters selected.

Add state to the search bar.
Create a function that will add in the search info into the API request.